Fix issues when devices get added/removed by storing the name
Previously the index was stored which changed every time you added or removed an audio device. Now the name of the device is stored (as device IDs don't seem to be consistent?) which only leads to issues when you rename the device you are currently using, otherwise it's resistant to changes to audio devices.